Unit tests for StakingWallet modifications (unique staker list)
